Q: Step6-2 Array Parser (#186)
* create error.js file and modified arrayParser.js and typeCheckers.js

create newly a file named error.js to validate type and string with
single quotation marks in the middle of strings. With that, rename
funtion in typeCheckers.js to getType from typeChecker and add function
from error.js which was newly created.

* creat 4 files and modified 1 file

Create stack.js for step6-2 to parsed arrays in an outer array and
modified arrayParser.js file to implement parsing arrays with stack.js.
With that, add getType.js to separate fundtion and leave only literals in spearators.js. Also create design_2.md.

* modify 3 files

add functions in error.js to check if input data exists and input data is
valid. With that, call those functions in tokenizer in
arrayParser.js.Also, add some comments in design_2.md.